Remove base Secrets Manager interface from AWS impl
