+++ aws_secrets_manager.py
@@ -1,7 +1,5 @@
-from data_parser import CSMConfig, CSMDefinitions
 import hashlib
 import pprint
-from abc import ABC, abstractmethod
 from json import dumps, loads
 from typing import Dict, List
 
@@ -11,67 +9,12 @@ from botocore.exceptions import ClientError
 import base_ccloud
 from api_key_manager import CCloudAPIKey, CCloudAPIKeyList
 from clusters import CCloudClusterList, CCloudEnvironmentList
+from data_parser import CSMConfig, CSMDefinitions
+from secrets_manager_interface import CSMSecret, CSMSecretsList
 from service_account import CCloudServiceAccountList
 
 pp = pprint.PrettyPrinter(indent=2)
 client = boto3.client("secretsmanager")
-
-
-class CSMSecret:
-    def __init__(
-        self,
-        secret_name: str,
-        secret_value: Dict[str, str],
-        env_id: str,
-        sa_id: str,
-        sa_name: str,
-        cluster_id: str,
-        api_key: str,
-    ) -> None:
-        self.secret_name = secret_name
-        self.secret_value = secret_value
-        self.env_id = env_id
-        self.sa_id = sa_id
-        self.sa_name = sa_name
-        self.cluster_id = cluster_id
-        self.api_key = api_key
-
-
-class CSMSecretsList(ABC):
-    secret: Dict[str, CSMSecret]
-
-    def __init__(self) -> None:
-        pass
-
-    def create_secret_name_string(
-        self, secret_name_prefix: str, seperator: str, env_id: str, cluster_id: str, sa_id: str
-    ):
-        # seperator = "/"
-        # secretName = env_name + sep + cluster_name + sep + sa_name
-        secret_name = (
-            (str(seperator + secret_name_prefix) if secret_name_prefix else "")
-            + seperator
-            + "ccloud"
-            + seperator
-            + sa_id
-            + seperator
-            + env_id
-            + seperator
-            + cluster_id
-        )
-        return secret_name
-
-    @abstractmethod
-    def read_all_secrets(self, filter: Dict[str, List[str]]):
-        pass
-
-    @abstractmethod
-    def find_secret(self, sa_name: str, sa_list: CCloudServiceAccountList, cluster_id: str = None) -> List[CSMSecret]:
-        pass
-
-    @abstractmethod
-    def create_or_update_secret():
-        pass
 
 
 class AWSSecret(CSMSecret):
@@ -228,61 +171,6 @@ class AWSSecretsList(CSMSecretsList):
         return
 
 
-# def create_secret_name(env_id: str, cluster_id: str, sa_id: str, api_key_name: str, secret_name_prefix: str):
-#     sep = "/"
-#     # secretName = env_name + sep + cluster_name + sep + sa_name
-#     secret_name = (
-#         (str(sep + secret_name_prefix) if secret_name_prefix else "")
-#         + sep
-#         + "ccloud"
-#         + sep
-#         + sa_id
-#         + sep
-#         + env_id
-#         + sep
-#         + cluster_id
-#     )
-#     return secret_name
-
-
-# def create_filter_tags(key_value: str, tags: dict):
-#     filter = [{"Key": key_value, "Values": [v]} for k, v in tags.items()]
-#     return filter
-
-
-# def render_secret_tags_format(tags: dict):
-#     return [{"Key": k, "Value": v} for k, v in tags.items()]
-
-
-# def create_digest(json_object_data):
-#     output = hashlib.md5(dumps(json_object_data, sort_keys=True).encode("utf-8")).hexdigest()
-#     return output
-
-
-# def list_secrets(key_value: str, filter_tags: dict):
-#     filter = create_filter_tags(key_value, filter_tags)
-#     resp = client.list_secrets(Filters=filter)
-#     if resp["ResponseMetadata"]["HTTPStatusCode"] != 200:
-#         raise Exception("AWS Secrets Manager List request failed. Please check the error and try again." + dumps(resp))
-#     return resp["SecretList"]
-
-
-# def get_secret(secret_name: str):
-#     try:
-#         resp = client.get_secret_value(SecretId=secret_name)
-#     except ClientError as e:
-#         if e.response["Error"]["Code"] == "ResourceNotFoundException":
-#             print("Secret Not Found.")
-#             return {}
-#         else:
-#             raise e
-#     if resp["ResponseMetadata"]["HTTPStatusCode"] != 200:
-#         raise Exception(
-#             "AWS Secrets Manager Get Secret request failed. Please check the error and try again." + dumps(resp)
-#         )
-#     return resp
-
-
 def add_secrets_to_rest_proxy_user(
     env_id: str,
     cluster_id: str,
@@ -342,47 +230,6 @@ def add_secrets_to_rest_proxy_user(
         )
         update_secret(rest_proxy_secret_name, loads(secret_data["SecretString"]), secret_string, secret_tags)
     return
-
-
-# def create_secret(secret_name: str, secret_values: dict, secret_tags: list):
-#     print("Trying to create a secret with the following details:")
-#     pp.pprint({"Secret Name": secret_name, "Secret Tags": secret_tags})
-#     resp = client.create_secret(
-#         Name=secret_name,
-#         Description="API Key & Secret generated by the CI/CD process.",
-#         SecretString=dumps(secret_values),
-#         Tags=secret_tags,
-#     )
-#     print("Secret Created Successfully. Secret Details as follows:")
-#     return resp
-
-
-# def update_secret(secret_name: str, old_secret_values: str, new_secret_values: dict, new_secret_tags: list):
-#     old_hash = create_digest(loads(old_secret_values))
-#     new_hash = create_digest(new_secret_values)
-#     if old_hash == new_hash:
-#         print("Not updating Secret with the provided value as current value is same as the older value.")
-#     else:
-#         print(
-#             "Updating "
-#             + secret_name
-#             + " with the new API Key & Secret values. API Key ID: "
-#             + new_secret_values["username"]
-#         )
-#         resp = client.put_secret_value(
-#             SecretId=secret_name,
-#             SecretString=dumps(new_secret_values),
-#         )
-#         print("Updated secret successfully with new API Key/Secret. Secret Details:")
-#         pp.pprint(resp)
-#     print("Adding/Updating Tags as follows:")
-#     pp.pprint(new_secret_tags)
-#     resp = client.tag_resource(SecretId=secret_name, Tags=new_secret_tags)
-#     if resp["ResponseMetadata"]["HTTPStatusCode"] != 200:
-#         print("Was not able to update the secret tags.")
-#     else:
-#         print("Tags Added successfully.")
-#     return
 
 
 def run_aws_sec_mgr_workflow(
@@ -551,7 +398,5 @@ if __name__ == "__main__":
     api_keys = {}
     with open("api_key_values.json", "r", encoding="utf-8") as input_file:
         api_keys = loads(input_file.read())
-    # run_aws_sec_mgr_workflow("get", api_keys, args.add_as_rest_proxy_user,
-    #                          False, args.secret_name_prefix)
     run_aws_sec_mgr_workflow("store", api_keys, args.add_as_rest_proxy_user, args.secret_name_prefix)
     print("")
